Split sdk class from index
