+++ src/README_HTML.md
@@ -6,7 +6,7 @@ So on these pages you'll find some very technical but super handy descriptions o
 
 ## Controllers
 
--   [Default methods (SDK)](classes/index.SDK.html)
+-   [Default methods (SDK)](classes/sdk.SDK.html)
 -   [ActionController](classes/src_controllers_ActionController.ActionController.html)
 -   [AnimationController](classes/src_controllers_AnimationController.AnimationController.html)
 -   [CanvasController](classes/src_controllers_CanvasController.CanvasController.html)
